+++ trollflow2/tests/test_launcher.py
@@ -24,6 +24,10 @@
 import unittest
 import yaml
 try:
+    from yaml import UnsafeLoader
+except ImportError:
+    from yaml import Loader as UnsafeLoader
+try:
     from unittest import mock
 except ImportError:
     import mock
@@ -31,6 +35,9 @@ except ImportError:
 yaml_test1 = """common:
   something: foo
   min_coverage: 5.0
+  subscribe_topics:
+    - /topic1
+    - /topic2
 product_list:
   euron1:
     areaname: euron1
@@ -112,7 +119,7 @@ class TestGetAreaPriorities(unittest.TestCase):
 
     def test_get_area_priorities(self):
         from trollflow2.launcher import get_area_priorities
-        prodlist = yaml.load(yaml_test1)
+        prodlist = yaml.load(yaml_test1, Loader=UnsafeLoader)
 
         priorities = get_area_priorities(prodlist)
         self.assertTrue(1 in priorities)
@@ -128,7 +135,7 @@ class TestMessageToJobs(unittest.TestCase):
 
     def test_message_to_jobs(self):
         from trollflow2.launcher import message_to_jobs
-        prodlist = yaml.load(yaml_test1)
+        prodlist = yaml.load(yaml_test1, Loader=UnsafeLoader)
         msg = mock.MagicMock()
         msg.data = {'uri': 'foo'}
 
@@ -152,7 +159,7 @@ class TestMessageToJobs(unittest.TestCase):
 
     def test_message_to_jobs_minimal(self):
         from trollflow2.launcher import message_to_jobs
-        prodlist = yaml.load(yaml_test_minimal)
+        prodlist = yaml.load(yaml_test_minimal, Loader=UnsafeLoader)
         msg = mock.MagicMock()
         msg.data = {'uri': 'foo'}
         jobs = message_to_jobs(msg, prodlist)
@@ -175,11 +182,15 @@ class TestMessageToJobs(unittest.TestCase):
 
 class TestRun(unittest.TestCase):
 
+    def setUp(self):
+        self.config = yaml.load(yaml_test1, Loader=UnsafeLoader)
+
+    @mock.patch('trollflow2.launcher.yaml.load')
+    @mock.patch('trollflow2.launcher.open')
     @mock.patch('trollflow2.launcher.process')
-    @mock.patch('trollflow2.launcher.time.sleep')
     @mock.patch('trollflow2.launcher.Process')
     @mock.patch('trollflow2.launcher.ListenerContainer')
-    def test_run(self, lc_, Process, sleep, process):
+    def test_run(self, lc_, Process, process, _open, yaml_load):
         from trollflow2.launcher import run
         listener = mock.MagicMock()
         listener.output_queue.get.return_value = 'foo'
@@ -187,28 +198,39 @@ class TestRun(unittest.TestCase):
         proc_ret = mock.MagicMock()
         Process.return_value = proc_ret
         # stop looping
-        sleep.side_effect = KeyboardInterrupt
+        proc_ret.join.side_effect = KeyboardInterrupt
+        yaml_load.return_value = self.config
         prod_list = 'bar'
-        topics = 'baz'
         try:
-            run(topics, prod_list)
+            run(prod_list)
         except KeyboardInterrupt:
             pass
         listener.output_queue.called_once()
         Process.assert_called_with(args=('foo', prod_list), target=process)
         proc_ret.start.assert_called_once()
         proc_ret.join.assert_called_once()
-        sleep.called_once_with(5)
+        lc_.assert_called_with(topics=['/topic1', '/topic2'])
+        # Subscriber topics are removed from config
+        self.assertTrue('subscribe_topics' not in self.config['common'])
+        # Topics are given as command line option
+        lc_.reset_mock()
+        try:
+            run(prod_list, topics=['/topic3'])
+        except KeyboardInterrupt:
+            pass
+        lc_.assert_called_with(topics=['/topic3'])
 
+    @mock.patch('trollflow2.launcher.yaml.load')
+    @mock.patch('trollflow2.launcher.open')
     @mock.patch('trollflow2.launcher.ListenerContainer')
-    def test_run_keyboard_interrupt(self, lc_):
+    def test_run_keyboard_interrupt(self, lc_, _open, yaml_load):
         from trollflow2.launcher import run
         listener = mock.MagicMock()
         get = mock.Mock()
         get.side_effect = KeyboardInterrupt
         listener.output_queue.get = get
         lc_.return_value = listener
-        run(0, 1)
+        run(0)
         listener.stop.assert_called_once()
 
 
@@ -223,16 +245,21 @@ class TestExpand(unittest.TestCase):
 
 class TestProcess(unittest.TestCase):
 
+    @mock.patch('trollflow2.launcher.traceback')
+    @mock.patch('trollflow2.launcher.sendmail')
     @mock.patch('trollflow2.launcher.expand')
     @mock.patch('trollflow2.launcher.yaml')
     @mock.patch('trollflow2.launcher.message_to_jobs')
     @mock.patch('trollflow2.launcher.open')
-    def test_process(self, open_, message_to_jobs, yaml, expand):
+    def test_process(self, open_, message_to_jobs, yaml_, expand, sendmail,
+                     traceback):
         from trollflow2.launcher import process
         fid = mock.MagicMock()
         fid.read.return_value = yaml_test1
         open_.return_value.__enter__.return_value = fid
-        yaml.load.return_value = "foo"
+        mock_config = mock.MagicMock()
+        yaml_.load.return_value = mock_config
+        yaml_.YAMLError = yaml.YAMLError
         fun1 = mock.MagicMock()
         # Return something resembling a config
         expand.return_value = {"workers": [{"fun": fun1}]}
@@ -240,13 +267,30 @@ class TestProcess(unittest.TestCase):
         message_to_jobs.return_value = {1: {"job1": dict([])}}
         process("msg", "prod_list")
         open_.assert_called_with("prod_list")
-        yaml.load.assert_called_once()
+        yaml_.load.assert_called_once()
         message_to_jobs.assert_called_with("msg", {"workers": [{"fun": fun1}]})
         fun1.assert_called_with({'job1': {}, 'processing_priority': 1})
         # Test that errors are propagated
-        yaml.load.side_effect = KeyboardInterrupt
+        fun1.side_effect = KeyboardInterrupt
         with self.assertRaises(KeyboardInterrupt):
             process("msg", "prod_list")
+        # Test crash hander call.  This will raise KeyError as there
+        # are no configured workers in the config returned by expand()
+        traceback.format_exc.return_value = 'baz'
+        crash_handlers = {"crash_handlers": {"config": {"foo": "bar"},
+                                             "handlers": [{"fun": sendmail}]}}
+        expand.return_value = crash_handlers
+        process("msg", "prod_list")
+        config = crash_handlers['crash_handlers']['config']
+        sendmail.assert_called_once_with(config, 'baz')
+
+        # Test failure in open(), e.g. a missing file
+        open_.side_effect = IOError
+        process("msg", "prod_list")
+
+        # Test failure in yaml.load(), e.g. bad formatting
+        open_.side_effect = yaml.YAMLError
+        process("msg", "prod_list")
 
 
 def suite():