+++ CHANGELOG.md
@@ -1,3 +1,61 @@
+## Version 0.4.1 (2019/04/10)
+
+
+### Pull Requests Merged
+
+#### Bugs fixed
+
+* [PR 33](https://github.com/pytroll/trollflow2/pull/33) - Fix hanging publisher
+
+#### Features added
+
+* [PR 32](https://github.com/pytroll/trollflow2/pull/32) - Feature sunlight coverage
+
+In this release 2 pull requests were closed.
+
+## Version 0.4.0 (2019/04/08)
+
+
+### Pull Requests Merged
+
+#### Bugs fixed
+
+* [PR 30](https://github.com/pytroll/trollflow2/pull/30) - Use only one sensor for coverage calculations
+* [PR 26](https://github.com/pytroll/trollflow2/pull/26) - Handle aliases for iterable metadata values
+
+#### Features added
+
+* [PR 30](https://github.com/pytroll/trollflow2/pull/30) - Use only one sensor for coverage calculations
+* [PR 27](https://github.com/pytroll/trollflow2/pull/27) - Add overviews to output images
+* [PR 26](https://github.com/pytroll/trollflow2/pull/26) - Handle aliases for iterable metadata values
+* [PR 25](https://github.com/pytroll/trollflow2/pull/25) - Add a possibility to send emails about crashes
+* [PR 24](https://github.com/pytroll/trollflow2/pull/24) - Check `collection_area_id` in the input metadata
+* [PR 23](https://github.com/pytroll/trollflow2/pull/23) - Add a possibility to define subscribe topics in config file
+* [PR 22](https://github.com/pytroll/trollflow2/pull/22) - Make publish topic composable
+* [PR 6](https://github.com/pytroll/trollflow2/pull/6) - Add a docker example
+
+In this release 10 pull requests were closed.
+
+## Version 0.3.0 (2019/03/19)
+
+
+### Pull Requests Merged
+
+#### Bugs fixed
+
+* [PR 20](https://github.com/pytroll/trollflow2/pull/20) - Handling nicer the situation where the scene cannot be created when th…
+* [PR 19](https://github.com/pytroll/trollflow2/pull/19) - Fix compatibility issues caused by changes introduced in pyyaml 5.1
+* [PR 18](https://github.com/pytroll/trollflow2/pull/18) - First take the info from the scene object, then update with what is p…
+* [PR 15](https://github.com/pytroll/trollflow2/pull/15) - Fix plist_iter to provide the area and product keys too
+
+#### Features added
+
+* [PR 20](https://github.com/pytroll/trollflow2/pull/20) - Handling nicer the situation where the scene cannot be created when th…
+* [PR 17](https://github.com/pytroll/trollflow2/pull/17) - Allow formats to be specified at any level
+* [PR 16](https://github.com/pytroll/trollflow2/pull/16) - Make it possible to delay composite creation
+
+In this release 7 pull requests were closed.
+
 ## Version 0.2.0 (2019/02/28)